requirements.txt
python-cowsay
Flask
